+++ packages/storage/src/internals/apis/getProperties.ts
@@ -17,10 +17,10 @@ import { GetPropertiesOutput } from '../types/outputs';
  *
  * @internal
  */
-export function getProperties(
+export const getProperties = (
 	input: GetPropertiesInput,
-): Promise<GetPropertiesOutput> {
-	return getPropertiesInternal(Amplify, {
+): Promise<GetPropertiesOutput> =>
+	getPropertiesInternal(Amplify, {
 		path: input.path,
 		options: {
 			useAccelerateEndpoint: input?.options?.useAccelerateEndpoint,
@@ -31,4 +31,3 @@ export function getProperties(
 		// Type casting is necessary because `getPropertiesInternal` supports both Gen1 and Gen2 signatures, but here
 		// given in input can only be Gen2 signature, the return can only ben Gen2 signature.
 	}) as Promise<GetPropertiesOutput>;
-}