+++ packages/storage/src/internals/index.ts
@@ -1,6 +1,7 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
+export { LocationCredentialsProvider } from '../providers/s3/types/options';
 export { StorageSubpathStrategy } from '../types/options';
 
 export { Permission } from './types/common';
@@ -14,8 +15,9 @@ export {
 	GetPropertiesInput,
 	GetUrlInput,
 	CopyInput,
-	ListInputWithPath,
+	ListInput,
 	RemoveInput,
+	UploadDataInput,
 	DownloadDataInput,
 } from './types/inputs';
 export {
@@ -24,7 +26,10 @@ export {
 	GetPropertiesOutput,
 	GetUrlOutput,
 	RemoveOutput,
+	UploadDataOutput,
 	DownloadDataOutput,
+	ListOutput,
+	CopyOutput,
 } from './types/outputs';
 
 export { getDataAccess } from './apis/getDataAccess';
@@ -33,7 +38,9 @@ export { list } from './apis/list';
 export { getProperties } from './apis/getProperties';
 export { getUrl } from './apis/getUrl';
 export { remove } from './apis/remove';
+export { uploadData } from './apis/uploadData';
 export { downloadData } from './apis/downloadData';
+export { copy } from './apis/copy';
 
 /*
 CredentialsStore exports
@@ -55,5 +62,4 @@ export {
 	GetLocationCredentialsInput,
 	GetLocationCredentialsOutput,
 } from './types/credentials';
-
 export { AWSTemporaryCredentials } from '../providers/s3/types/options';