+++ packages/storage/src/providers/s3/apis/internal/uploadData/multipart/initialUpload.ts
@@ -3,12 +3,15 @@
 
 import { StorageAccessLevel } from '@aws-amplify/core';
 
-import { ContentDisposition, ResolvedS3Config } from '../../../types/options';
-import { StorageUploadDataPayload } from '../../../../../types';
-import { Part, createMultipartUpload } from '../../../utils/client/s3data';
-import { logger } from '../../../../../utils';
-import { calculateContentCRC32 } from '../../../utils/crc32';
-import { constructContentDisposition } from '../../../utils/constructContentDisposition';
+import {
+	ContentDisposition,
+	ResolvedS3Config,
+} from '../../../../types/options';
+import { StorageUploadDataPayload } from '../../../../../../types';
+import { Part, createMultipartUpload } from '../../../../utils/client/s3data';
+import { logger } from '../../../../../../utils';
+import { calculateContentCRC32 } from '../../../../utils/crc32';
+import { constructContentDisposition } from '../../../../utils/constructContentDisposition';
 
 import {
 	cacheMultipartUpload,