+++ packages/aws-amplify/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [6.6.3](https://github.com/aws-amplify/amplify-js/compare/aws-amplify@6.6.2...aws-amplify@6.6.3) (2024-09-30)
+
+**Note:** Version bump only for package aws-amplify
+
 ## [6.6.2](https://github.com/aws-amplify/amplify-js/compare/aws-amplify@6.6.1...aws-amplify@6.6.2) (2024-09-17)
 
 **Note:** Version bump only for package aws-amplify