+++ packages/geo/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.0.50](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/geo@3.0.49...@aws-amplify/geo@3.0.50) (2024-09-30)
+
+**Note:** Version bump only for package @aws-amplify/geo
+
 ## [3.0.49](https://github.com/aws-amplify/amplify-js/compare/@aws-amplify/geo@3.0.48...@aws-amplify/geo@3.0.49) (2024-09-17)
 
 **Note:** Version bump only for package @aws-amplify/geo